+++ app/components/bulmacomp/navbar_component.rb
@@ -40,13 +40,19 @@ module Bulmacomp
     #   to generate content
     # @option opts [String] :brand
     #   navbar brand content
+    # @option opts [String] :navbar_start
+    #   optionally navbar-start content
+    # @option opts [String] :navbar_end
+    #   optionally navbar-end content
     # @option opts [String] :*
     #   each other key going as tag option, default is
     #   class: 'navbar', aria_label: 'main navigation', aria_role: 'navigation'
     # @yield [optional] navbar content
-    def initialize(brand: nil, **opts)
+    def initialize(brand: nil, navbar_start: nil, navbar_end: nil, **opts)
       super
       @brand = brand
+      @navbar_start = tag.div navbar_start, class: 'navbar-start' if navbar_start.present?
+      @navbar_end = tag.div navbar_end, class: 'navbar-end' if navbar_end.present?
       @opts = { class: 'navbar', aria: { role: 'navigation', label: 'main navigation' } }.merge(opts)
     end
 
@@ -65,10 +71,10 @@ module Bulmacomp
 
     # @return [String] html_safe navbar-menu tag
     def navbar_menu
-      tag.div(content, class: 'navbar-menu', data: {target: 'page-menu'})
+      tag.div(safe_join([@navbar_start, content, @navbar_end]), class: 'navbar-menu', data: { target: 'page-menu' })
     end
 
-    # return [String] html_safe navbar-burger link
+    # @return [String] html_safe navbar-burger link
     def navbar_burger
       tag.a tag.span(aria_hidden: true) * 3, class: 'navbar-burger', aria_label: 'menu', aria_expanded: 'false',
                                              data_action: 'page#toggleMenu'