+++ CHANGELOG.md
@@ -3,11 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-### [1.0.4]
-## Fixed
+## [1.1.0]
+### Added
+* new option for navbar component:
+* * :navbar-start for popolate `navbar-start` div in `navbar-menu`
+* * :navbar-end for popolate `navbar-end` div in `navbar-menu`
+
+## [1.0.4]
+### Fixed
 * fixed data target required to open menu
 
-### [1.0.3]
+## [1.0.3]
 ### Fixed
 * rendering navbar component fixed
 ### Added